Refactor event handler
